% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRegions.R
\name{getRegions}
\alias{getRegions}
\title{getRegions}
\usage{
getRegions(lats, longs)
}
\arguments{
\item{lats}{Latitude values in decimal degrees}

\item{longs}{Longitude values in decimal degrees}
}
\value{
Returns a data frame of ECOREGION, PROVINCE, and REALMs for each point

seealso \code{\link{sp::over}}
}
\description{
\code{getRegions} Gets regional information for your dataset
}
\examples{
data(regions)

latlong <- data.frame(lat = c(50.0944444, 33.4691667, 34.0348833,
     55.0869876, 51.7787776,
     49.6088889, -35.750729),
     long=c(-127.55889, -119.53028,
     -119.70308, -161.87444,
     178.30199, -126.78056, 150.259155))

regionalData <- getRegions(latlong$lat, latlong$long)

newdata <- cbind(latlong, regionalData)

newdata
}
\author{
Jarrett Byrnes.
}

