% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeMEOWmap.R
\name{makeMEOWmap}
\alias{makeMEOWmap}
\title{makeMEOWmap}
\usage{
makeMEOWmap(newdata, fillColName, regionColName = type, type = "ECOREGION",
  fillPal = brewer.pal(11, "Spectral"), pal = "Spectral",
  pathCol = "black", pathAlpha = 1, guide = guide_colourbar(title =
  fillColName), dataOut = FALSE, na.value = NA, add.worldmap = FALSE,
  fillAlphaColName = NULL, excludeNoDataAreas = T, prevggplot = NULL, ...)
}
\arguments{
\item{newdata}{dataset to be used}

\item{fillColName}{name of the column with information to be used for a fill value}

\item{regionColName}{data name of column with ecoregion, province, or realm names}

\item{type}{are these values from an ECOREGION, PROVINCE, or REALM map. Defaults to "ECOREGION". Note all caps.}

\item{fillPal}{The palatte used by \code{\link{scale_fill_gradientn}}. Defaults to \code{\link{brewer.pal}(11, "Spectral")}}

\item{pal}{The palatte used by \code{\link{scale_fill_manual}} if data is categorical. Defaults to "spectral" from \code{\link{RColorBrewer}}}

\item{pathCol}{The path color for regional outlines. Defaults to "black"}

\item{pathAlpha}{The alpha of the regional outlines. Defaults to 1.}

\item{guide}{What kind of guide should be used, and what should it's title be? Defaults to \code{\link{guide_colourbar}(title=fillColName)}}

\item{dataOut}{Return a merged dataframe for plotting instead of a plot? Defaults to \code{FALSE}}

\item{na.value}{How are NAs areas handled for fill? Defaults to making them not plot - NA.}

\item{add.worldmap}{Should a map of the world be plotted under the regions? Defaults to FALSE}

\item{fillAlphaColName}{Colname of column used to scale alpha level of fill}

\item{excludeNoDataAreas}{Exclude areas/regions from the plot with no data?}

\item{prevggplot}{A ggplot argument that this plot will be added on top of. If null, a new
ggplot object is created.}

\item{...}{Other arguments to be supplied to color scale}
}
\value{
Returns \code{\link{ggplot}} object
}
\description{
\code{makeMEOWmap} takes a dataset with values and a MEOW column and makes a map in \code{\link{ggplot2}}
}
\details{
\code{makeMEOWmap} takes a dataset with values and a MEOW column and makes a map in \code{\link{ggplot2}}
}
\examples{
data(regions.df)

ndf <- data.frame(Ecoregions = levels(regions.df$ECOREGION),
Values = runif(length(levels(regions.df$ECOREGION)), 0,100))

makeMEOWmap(ndf, fillColName="Values", regionColName="Ecoregions")
}
\author{
Jarrett Byrnes.
}

