% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meowR.R
\docType{package}
\name{meowR}
\alias{meowR}
\alias{meowR-package}
\title{meowR}
\description{
Methods and data for adding Spalding et al.’s Marine Ecoregions of the World to plots. Data from http://www.marineregions.org/downloads.php downloaded on June 1st 2015.
}
\author{
Jarrett Byrnes
}

