% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meowR.R
\docType{data}
\name{regions}
\alias{provinces}
\alias{provinces.df}
\alias{realms}
\alias{realms.df}
\alias{regions}
\alias{regions.df}
\title{Marine Ecoregions of the World Data}
\usage{
data(regions)
data(provinces)
data(realms)

data(regions.df)
data(provinces.df)
data(realms.df)
}
\description{
Data from \link{http://www.marineregions.org/downloads.php}. All datasets provides either as \code{\link{SpatialPolygonsDataFrame}} objects or, with .df, as \code{\link{data.frame}} objects
}
\examples{

data(regions)
data(provinces)
data(realms)

par(mfrow=c(2,2), mar=c(0,1,2,0))
plot(regions, main="Ecoregion")
plot(provinces, lwd=2, border="red", main="Province")
plot(realms, lwd=2, border="blue", main="Realm")
par(mfrow=c(1,1), mar=c(5, 4, 4, 2) + 0.1)

#or with ggplot2
base_province_ggplot <- ggplot(provinces.df) + theme_bw() +
aes(long,lat,group=group) + 
  geom_polygon(fill=NA) +
  geom_path(color="black") +
  coord_equal() 
  
  base_province_ggplot
}
\references{
Spalding, M.D., Fox, H.E., Allen, G.R., Davidson, N., Ferdaña, Z.A., Finlayson, M., Halpern, B.S., Jorge, M.A., Lombana, A.L., Lourie, S.A., 2007. Marine ecoregions of the world: a bioregionalization of coastal and shelf areas. Bioscience 57, 573–583.
}
\keyword{data}

