% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meowR.R
\docType{data}
\name{worldmap}
\alias{worldmap}
\alias{worldmap.df}
\title{Map of the world}
\usage{
data(worldmap)
data(worldmap.df)
}
\description{
Data from \link{maps} package. All datasets provides either as \code{\link{SpatialPolygonsDataFrame}} objects or, with .df, as \code{\link{data.frame}} objects
}
\examples{

data(worldmap)
plot(worldmap, type="l")

data(worldmap.df)
ggplot(worldmap.df, aes(x=longitude, y=latitude))+geom_path()

}
\keyword{data}

